+++ Robust.Client/Console/Commands/Debug.cs
@@ -427,7 +427,7 @@ namespace Robust.Client.Console.Commands
 
         public bool Execute(IDebugConsole console, params string[] args)
         {
-            var window = new SS14Window("UITest");
+            var window = new SS14Window();
             var tabContainer = new TabContainer();
             window.Contents.AddChild(tabContainer);
             var scroll = new ScrollContainer();
