+++ Robust.Client/Interfaces/UserInterface/IUserInterfaceManager.cs
@@ -99,6 +99,9 @@ namespace Robust.Client.Interfaces.UserInterface
         void RemoveModal(Control modal);
 
         void Render(IRenderHandle renderHandle);
+
+        void QueueStyleUpdate(Control control);
+        void QueueLayoutUpdate(Control control);
     }
 }
