+++ Robust.Client/UserInterface/Controls/BaseButton.cs
@@ -16,14 +16,6 @@ namespace Robust.Client.UserInterface.Controls
         private bool _disabled;
         private bool _pressed;
 
-        protected BaseButton()
-        {
-        }
-
-        protected BaseButton(string name) : base(name)
-        {
-        }
-
         /// <summary>
         ///     Controls mode of operation in relation to press/release events.
         /// </summary>
@@ -233,7 +225,7 @@ namespace Robust.Client.UserInterface.Controls
             Normal = 0,
             Pressed = 1,
             Hover = 2,
-            Disabled = 3,
+            Disabled = 3
         }
 
         public class ButtonEventArgs : EventArgs
@@ -279,7 +271,7 @@ namespace Robust.Client.UserInterface.Controls
             ///     <see cref="BaseButton.OnPressed"/> fires when the mouse button causing them is released.
             ///     This is the default and most intuitive method.
             /// </summary>
-            Release = 1,
+            Release = 1
         }
     }
 }
