+++ Robust.Client/UserInterface/Controls/CheckButton.cs
@@ -2,8 +2,5 @@ namespace Robust.Client.UserInterface.Controls
 {
     public class CheckButton : Button
     {
-        public CheckButton()
-        {
-        }
     }
 }
