+++ Robust.Client/UserInterface/Controls/HBoxContainer.cs
@@ -5,14 +5,6 @@
     /// </summary>
     public class HBoxContainer : BoxContainer
     {
-        public HBoxContainer() : base()
-        {
-        }
-
-        public HBoxContainer(string name) : base(name)
-        {
-        }
-
         private protected override bool Vertical => false;
     }
 }
