+++ Robust.Client/UserInterface/Controls/VBoxContainer.cs
@@ -5,14 +5,6 @@
     /// </summary>
     public class VBoxContainer : BoxContainer
     {
-        public VBoxContainer()
-        {
-        }
-
-        public VBoxContainer(string name) : base(name)
-        {
-        }
-
         private protected override bool Vertical => true;
     }
 }
