+++ Robust.Client/UserInterface/Controls/VScrollBar.cs
@@ -4,6 +4,7 @@
     {
         public VScrollBar() : base(OrientationMode.Vertical)
         {
+            GrowHorizontal = GrowDirection.Begin;
         }
     }
 }
