+++ Robust.Client/UserInterface/Controls/VSplitContainer.cs
@@ -3,7 +3,5 @@ namespace Robust.Client.UserInterface.Controls
     public class VSplitContainer : SplitContainer
     {
         private protected sealed override bool Vertical => true;
-
-        public VSplitContainer() {}
     }
 }
