+++ Robust.Client/ViewVariables/ViewVariablesManager.cs
@@ -189,7 +189,7 @@ namespace Robust.Client.ViewVariables
                 instance = new ViewVariablesInstanceObject(this, _resourceCache);
             }
 
-            var window = new SS14Window("VV") {Title = "View Variables"};
+            var window = new SS14Window {Title = "View Variables"};
             instance.Initialize(window, obj);
             window.OnClose += () => _closeInstance(instance, false);
             _windows.Add(instance, window);
@@ -198,7 +198,7 @@ namespace Robust.Client.ViewVariables
 
         public async void OpenVV(ViewVariablesObjectSelector selector)
         {
-            var window = new SS14Window("VV") {Title = "View Variables"};
+            var window = new SS14Window {Title = "View Variables"};
             var loadingLabel = new Label {Text = "Retrieving remote object data from server..."};
             window.Contents.AddChild(loadingLabel);
