+++ Robust.UnitTesting/Client/UserInterface/Controls/BoxContainerTest.cs
@@ -1,13 +1,7 @@
 using NUnit.Framework;
-using Robust.Client.Interfaces.UserInterface;
 using Robust.Client.UserInterface;
 using Robust.Client.UserInterface.Controls;
-using Robust.Client.Utility;
-using Robust.Shared.Interfaces.Resources;
-using Robust.Shared.IoC;
-using Robust.Shared.Utility;
 using Robust.Shared.Maths;
-
 
 namespace Robust.UnitTesting.Client.UserInterface.Controls
 {
@@ -21,11 +15,13 @@ namespace Robust.UnitTesting.Client.UserInterface.Controls
         public void TestLayoutBasic()
         {
             var boxContainer = new VBoxContainer {CustomMinimumSize = (50, 60)};
-            var control1 = new Control("1") {CustomMinimumSize = (20, 20)};
-            var control2 = new Control("2") {CustomMinimumSize = (30, 30)};
+            var control1 = new Control {CustomMinimumSize = (20, 20)};
+            var control2 = new Control {CustomMinimumSize = (30, 30)};
 
             boxContainer.AddChild(control1);
             boxContainer.AddChild(control2);
+
+            boxContainer.ForceRunLayoutUpdate();
 
             Assert.That(control1.Position, Is.EqualTo(Vector2.Zero));
             Assert.That(control1.Size, Is.EqualTo(new Vector2(50, 20)));
@@ -39,14 +35,16 @@ namespace Robust.UnitTesting.Client.UserInterface.Controls
         public void TestLayoutExpand()
         {
             var boxContainer = new VBoxContainer {CustomMinimumSize = (50, 60)};
-            var control1 = new Control("1")
+            var control1 = new Control
             {
                 SizeFlagsVertical = Control.SizeFlags.FillExpand
             };
-            var control2 = new Control("2") {CustomMinimumSize = (30, 30)};
+            var control2 = new Control {CustomMinimumSize = (30, 30)};
 
             boxContainer.AddChild(control1);
             boxContainer.AddChild(control2);
+
+            boxContainer.ForceRunLayoutUpdate();
 
             Assert.That(control1.Position, Is.EqualTo(Vector2.Zero));
             Assert.That(control1.Size, Is.EqualTo(new Vector2(50, 29)));
@@ -60,11 +58,11 @@ namespace Robust.UnitTesting.Client.UserInterface.Controls
         public void TestCalcMinSize()
         {
             var boxContainer = new VBoxContainer();
-            var control1 = new Control("1")
+            var control1 = new Control
             {
                 CustomMinimumSize = (50, 30)
             };
-            var control2 = new Control("2") {CustomMinimumSize = (30, 50)};
+            var control2 = new Control {CustomMinimumSize = (30, 50)};
 
             boxContainer.AddChild(control1);
             boxContainer.AddChild(control2);
@@ -76,19 +74,21 @@ namespace Robust.UnitTesting.Client.UserInterface.Controls
         public void TestTwoExpand()
         {
             var boxContainer = new VBoxContainer {CustomMinimumSize = (30, 82)};
-            var control1 = new Control("1")
+            var control1 = new Control
             {
                 SizeFlagsVertical = Control.SizeFlags.FillExpand
             };
-            var control2 = new Control("2")
+            var control2 = new Control
             {
                 SizeFlagsVertical = Control.SizeFlags.FillExpand
             };
-            var control3 = new Control("3") {CustomMinimumSize = (0, 50)};
+            var control3 = new Control {CustomMinimumSize = (0, 50)};
 
             boxContainer.AddChild(control1);
             boxContainer.AddChild(control3);
             boxContainer.AddChild(control2);
+
+            boxContainer.ForceRunLayoutUpdate();
 
             Assert.That(control1.Position, Is.EqualTo(Vector2.Zero));
             Assert.That(control1.Size, Is.EqualTo(new Vector2(30, 15)));
