+++ Robust.UnitTesting/Client/UserInterface/Controls/CenterContainerTest.cs
@@ -23,6 +23,9 @@ namespace Robust.UnitTesting.Client.UserInterface.Controls
             var child = new Control {CustomMinimumSize = (50, 50)};
 
             container.AddChild(child);
+
+            container.ForceRunLayoutUpdate();
+
             Assert.That(container.CombinedMinimumSize, Is.EqualTo(new Vector2(50, 50)));
             Assert.That(child.Position, Is.EqualTo(new Vector2(25, 25)));
             Assert.That(child.Size, Is.EqualTo(new Vector2(50, 50)));
