+++ Robust.UnitTesting/Client/UserInterface/UserInterfaceManagerTest.cs
@@ -137,8 +137,8 @@ namespace Robust.UnitTesting.Client.UserInterface
         public void TestGrabKeyboardFocus()
         {
             Assert.That(_userInterfaceManager.KeyboardFocused, Is.Null);
-            var control1 = new Control("Control1") {CanKeyboardFocus = true};
-            var control2 = new Control("Control2") {CanKeyboardFocus = true};
+            var control1 = new Control {CanKeyboardFocus = true};
+            var control2 = new Control {CanKeyboardFocus = true};
 
             control1.GrabKeyboardFocus();
             Assert.That(_userInterfaceManager.KeyboardFocused, Is.EqualTo(control1));
@@ -155,8 +155,8 @@ namespace Robust.UnitTesting.Client.UserInterface
         public void TestGrabKeyboardFocusSteal()
         {
             Assert.That(_userInterfaceManager.KeyboardFocused, Is.Null);
-            var control1 = new Control("Control1") {CanKeyboardFocus = true};
-            var control2 = new Control("Control2") {CanKeyboardFocus = true};
+            var control1 = new Control {CanKeyboardFocus = true};
+            var control2 = new Control {CanKeyboardFocus = true};
 
             control1.GrabKeyboardFocus();
             control2.GrabKeyboardFocus();
@@ -172,8 +172,8 @@ namespace Robust.UnitTesting.Client.UserInterface
         public void TestGrabKeyboardFocusOtherRelease()
         {
             Assert.That(_userInterfaceManager.KeyboardFocused, Is.Null);
-            var control1 = new Control("Control1") {CanKeyboardFocus = true};
-            var control2 = new Control("Control2") {CanKeyboardFocus = true};
+            var control1 = new Control {CanKeyboardFocus = true};
+            var control2 = new Control {CanKeyboardFocus = true};
 
             control1.GrabKeyboardFocus();
             control2.ReleaseKeyboardFocus();
@@ -211,6 +211,8 @@ namespace Robust.UnitTesting.Client.UserInterface
             };
 
             _userInterfaceManager.RootControl.AddChild(control);
+
+            _userInterfaceManager.RootControl.ForceRunLayoutUpdate();
 
             var mouseEvent = new MouseButtonEventArgs(Mouse.Button.Left, false, Mouse.ButtonMask.None,
                 new Vector2(30, 30), false, false, false, false);
